Moved memory stick mode to InternalPreferences
