+++ src/main/java/org/jabref/preferences/GeneralPreferences.java
@@ -1,8 +1,6 @@
 package org.jabref.preferences;
 
-import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
 import org.jabref.logic.l10n.Language;
@@ -11,15 +9,11 @@ import org.jabref.model.database.BibDatabaseMode;
 public class GeneralPreferences {
     private final ObjectProperty<Language> language;
     private final ObjectProperty<BibDatabaseMode> defaultBibDatabaseMode;
-    private final BooleanProperty memoryStickMode;
 
     public GeneralPreferences(Language language,
-                              BibDatabaseMode defaultBibDatabaseMode,
-                              boolean memoryStickMode) {
+                              BibDatabaseMode defaultBibDatabaseMode) {
         this.language = new SimpleObjectProperty<>(language);
         this.defaultBibDatabaseMode = new SimpleObjectProperty<>(defaultBibDatabaseMode);
-
-        this.memoryStickMode = new SimpleBooleanProperty(memoryStickMode);
     }
 
     public Language getLanguage() {
@@ -44,17 +38,5 @@ public class GeneralPreferences {
 
     public void setDefaultBibDatabaseMode(BibDatabaseMode defaultBibDatabaseMode) {
         this.defaultBibDatabaseMode.set(defaultBibDatabaseMode);
-    }
-
-    public boolean isMemoryStickMode() {
-        return memoryStickMode.get();
-    }
-
-    public BooleanProperty memoryStickModeProperty() {
-        return memoryStickMode;
-    }
-
-    public void setMemoryStickMode(boolean memoryStickMode) {
-        this.memoryStickMode.set(memoryStickMode);
     }
 }
