+++ src/main/java/org/jabref/preferences/InternalPreferences.java
@@ -2,7 +2,9 @@ package org.jabref.preferences;
 
 import java.nio.file.Path;
 
+import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
@@ -14,11 +16,16 @@ public class InternalPreferences {
     private final ObjectProperty<Version> ignoredVersion;
     private final ObjectProperty<Path> lastPreferencesExportPath;
     private final StringProperty user;
+    private final BooleanProperty memoryStickMode;
 
-    public InternalPreferences(Version ignoredVersion, Path exportPath, String user) {
+    public InternalPreferences(Version ignoredVersion,
+                               Path exportPath,
+                               String user,
+                               boolean memoryStickMode) {
         this.ignoredVersion = new SimpleObjectProperty<>(ignoredVersion);
         this.lastPreferencesExportPath = new SimpleObjectProperty<>(exportPath);
         this.user = new SimpleStringProperty(user);
+        this.memoryStickMode = new SimpleBooleanProperty(memoryStickMode);
     }
 
     public Version getIgnoredVersion() {
@@ -47,5 +54,17 @@ public class InternalPreferences {
 
     public String getUser() {
         return user.get();
+    }
+
+    public boolean isMemoryStickMode() {
+        return memoryStickMode.get();
+    }
+
+    public BooleanProperty memoryStickModeProperty() {
+        return memoryStickMode;
+    }
+
+    public void setMemoryStickMode(boolean memoryStickMode) {
+        this.memoryStickMode.set(memoryStickMode);
     }
 }
